requirements.txt
boto3==1.34.154
Django==4.2.16
djangorestframework==3.15.2
openai==0.28.0
pdfplumber==0.9.0
pydub==0.25.1
python-dotenv==1.0.1
Requests==2.32.3
ffmpeg-python==0.2.0
librosa==0.9.2
soundfile==0.12.1
tqdm==4.66.1
mido==1.3.0
mir_eval==0.7
numpy==1.24.1
django-cors-headers==3.14.0
django-storages==1.13.2
djangorestframework-simplejwt==5.3.1